+++ asyncgit/src/sync/staging/stage_tracked.rs
@@ -97,7 +97,7 @@ mod test {
 		)
 		.unwrap();
 
-		let diff = get_diff(path, "test.txt", true).unwrap();
+		let diff = get_diff(path, "test.txt", true, None).unwrap();
 
 		assert_eq!(diff.lines, 3);
 		assert_eq!(
@@ -139,7 +139,7 @@ c = 4";
 		)
 		.unwrap();
 
-		let diff = get_diff(path, "test.txt", true).unwrap();
+		let diff = get_diff(path, "test.txt", true, None).unwrap();
 
 		assert_eq!(diff.lines, 5);
 		assert_eq!(
@@ -172,7 +172,8 @@ c = 4";
 
 		assert_eq!(get_statuses(path), (0, 1));
 
-		let diff_before = get_diff(path, "test.txt", true).unwrap();
+		let diff_before =
+			get_diff(path, "test.txt", true, None).unwrap();
 
 		assert_eq!(diff_before.lines, 5);
 
@@ -189,7 +190,7 @@ c = 4";
 
 		assert_eq!(get_statuses(path), (1, 1));
 
-		let diff = get_diff(path, "test.txt", true).unwrap();
+		let diff = get_diff(path, "test.txt", true, None).unwrap();
 
 		assert_eq!(diff.lines, 4);
 	}